move dtos around
