+++ backend/Bartender.Domain/DTO/Business/UpsertBusinessDto.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Bartender.Domain.DTO;
+﻿namespace Bartender.Domain.DTO.Business;
 
 public class UpsertBusinessDto
 {
