+++ backend/Bartender.Domain/DTO/Order/OrderItemsDto.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace Bartender.Domain.DTO.Orders;
+namespace Bartender.Domain.DTO.Order;
 
 public class OrderItemsDto
 {
