+++ backend/Bartender.Domain/DTO/Order/UpdateOrderStatusDto.cs
@@ -2,7 +2,7 @@
 using Bartender.Data.Enums;
 using System.ComponentModel.DataAnnotations;
 
-namespace Bartender.Domain.DTO.Orders;
+namespace Bartender.Domain.DTO.Order;
 
 public class UpdateOrderStatusDto
 {
