+++ backend/Bartender.Domain/DTO/Order/UpsertOrderDto.cs
@@ -1,7 +1,7 @@
 ﻿using Bartender.Data.Enums;
 using System.ComponentModel.DataAnnotations;
 
-namespace Bartender.Domain.DTO.Orders;
+namespace Bartender.Domain.DTO.Order;
 
 public class UpsertOrderDto
 {
