+++ backend/Bartender.Domain/DTO/Product/ProductDTO.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace Bartender.Domain.DTO.Products;
+namespace Bartender.Domain.DTO.Product;
 
 public class ProductDto
 {
