+++ backend/Bartender.Domain/DTO/ServiceResult.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Bartender.Domain;
+﻿namespace Bartender.Domain.DTO;
 public enum ErrorType
 {
     NotFound,
